反重力移動ユニット
反重力移動を別クラスに分け，壁からの重力の記述を加えました．
デバッグはしてません．

+++ group10/FixedPointer.java
@@ -1,9 +1,8 @@
 package group10;
-import robocode.*;
 
 public class FixedPointer extends AntiGrav{
 
-	public FixedPointer(int inx, int iny, int inweight){
+	public FixedPointer(double inx, double iny, double inweight){
 		x = inx;
 		y = iny;
 		weight = inweight;
